requirements.txt
numpy>=1.24
pyyaml>=6.0
torch>=2.0
tqdm>=4.66
